+++ routes/api/contacts.js
@@ -1,54 +1,18 @@
 import { Router } from 'express';
 import * as contactServises from '../../models/contacts.js';
+import * as controllersContact from '../../controlers/controllersContacts.js';
 const router = Router();
 
-router.get('/', async (req, res, next) => {
-  const result = res.json(await contactServises.listContacts());
-  // try {
-  // if (!result) {
-  //   return res.status(404).json({ message: 'Not found' });
-  // }
+router.get('/', controllersContact.getAll);
 
-  return result;
-  // } catch (error) {
-  //   res.status(500).json({ message: 'Server error' });
-  // }
-});
+router.get('/:contactId', controllersContact.getById);
 
-router.get('/:contactId', async (req, res, next) => {
-  // try {
-  const id = req.url.substring(1);
-  const result = res.json(await contactServises.getContactById(id));
-  console.log('result', result);
-  // if (!result) {
-  //   return res.status(404).json({ message: 'Not found' });
-  // }
-  return result;
-  // } catch (error) {
-  //   res.status(500).json({ message: 'Server error' });
-  // }
-});
+router.delete('/:contactId', controllersContact.deleteById);
 
-router.delete('/:contactId', async (req, res, next) => {
-  // try {
-  const id = req.url.substring(1);
-  res.json(await contactServises.removeContact(id));
-  // } catch (error) {
-  //   res.status(500).json({ message: 'Server error' });
-  // }
-});
-
-router.post('/', async (req, res, next) => {
-  // try {
-
-  res.json(await contactServises.addContact(req.body));
-  // } catch (error) {
-  //   res.status(500).json({ message: 'Server error' });
-  // }
-});
+router.post('/', controllersContact.add);
 
 router.put('/:contactId', async (req, res, next) => {
-  res.json({ message: 'template message' });
+  res.json(await contactServises.updateContact(id, reg.body));
 });
 
 export default router;
